ASoC: meson: aiu: refactor I2S encoder to support formatters
AIU I2S encoder controls clocks which are used for both audio output and
input. While the former is already supported in AIU, trying to add support
for the latter (AUDIN) creates a depedency between the 2 subsystems.

The workaround being proposed in this commit is to follow the same
approach used for Meson AXG platform:
- use the I2S DAI to only control clocks;
- create separate code for data formatters;
- connect formatters to the DAI at runtime through DAPM.

This commit introduces this support for AIU. Nothing changes from AIU's
features point of view, but it allows future introduction of AUDIN
support in the future.

Signed-off-by: Valerio Setti <vsetti@baylibre.com>