ASoC: meson: add support for I2S formatter in GX platform
Introduce the very basic support for an I2S formatter. This is basically
a copy of what's already in place for the AXG platform. From a practical
point of view nothing changes because the aiu-formatter is not doing
anything yet. A part from some boilerplate code change, the aiu-formatter
simply provides its IN/OUT widgets which will be used by DAPM to place
the aiu-formatter between the "AIU Encoder FIFO" and
"AIU Encoder Playback".

Data structures and functions were kept as close as possible to the ones
found in the AXG platform so that it will be easier to merge the 2
implementations in the future.

Signed-off-by: Valerio Setti <vsetti@baylibre.com>