ASoC: meson: add AUDIN formatter
This is the counterpart of aiu-formatter, but for the capturing part.
It basically takes care of the formatting management for I2S input.
Most of the clock configuration is handled in the aiu-encoder-i2s
(the same as for the aiu-formatter).

Unfortunately audin-formatter also does a bit of input clock configuration
because registers which are meant for this scope belong to its regmap
area and they are not accessible from the aiu-encoder-i2s component.

Signed-off-by: Valerio Setti <vsetti@baylibre.com>