ASoC: meson: aiu-encoder-i2s: start using gx-formatter
Add some new ops to the I2S DAI driver to let it start interacting with
the gx-formatter. The addition is quite simple: on probe a new gx_stream
is allocated to the capture/playback stream and this gets freed on remove.

Trigger is used to start/stop the formatter. For the time being it doesn't
perform any action because there is no formatter registered for streams,
but they will be added in follow-up commits.

Signed-off-by: Valerio Setti <vsetti@baylibre.com>