ASoC: meson: gx: add gx-formatter and gx-interface
These files are the basic block which allow to shape I2S in GX devices
the same as the AXG ones: the DAI backend only controls the interface
(i.e. clocks and pins) whereas a formatter takes care of properly
formatting the data.

gx-formatter and gx-interface  are stronly inspired to axg-tdm-formatter
and axg-tdm, respectively, and they keep only what's strictly necessary
for GX platform.
There is only a minor addition, "gx_formatter_create()", which is required
in order to let already existing AIU code to make use of this formatter
without making any devicetree change.

Signed-off-by: Valerio Setti <vsetti@baylibre.com>