ASoC: meson: aiu: transition to gx-formatter
Introduce "aiu-formatter-i2c" which implement the encoding of I2S playback
data. This is not a standalone driver but it has been integrated in the
already existing AIU driver in order to minimize changes and prevent any
devicetree modification.
A new DAPM widget has been added to the I2S playback stream to let
"aiu-formatter-i2s" power up/down when the stream is started/stopped,
respectively. This is where the formatter attaches to the backend DAI.

"aiu-encoder-i2s" now takes care of managing interface clocks, since all
the formatting part has been moved to "aiu-formatter-i2s".

Signed-off-by: Valerio Setti <vsetti@baylibre.com>