ASoC: meson: gx: add AUDIN FIFO driver
Add support for the frontend DAI of the capture interface which is in
charge of receiving decoded data from "audin-decoder-i2s" and move them to
an internal FIFO before they are block-transferring to RAM.

This component could ideally handle multiple input sources (SPDIF, I2S,
PCM, HDMI), but for the time being only I2S is added and has been tested.

Even though the platform ideally supports an interrupt associated to FIFO
fill level, this is not reliable. Indeed the interrupt triggering point
needs to be updated every time it triggers, leading to potential
synchronization issues in case of high system load.
Therefore an hrtimer is used in place of the interrupt to periodically
report received data.

Signed-off-by: Valerio Setti <vsetti@baylibre.com>